feat(an): add normalize web form, donation and pressure
